+++ sources/exec/exec.c
@@ -12,25 +12,6 @@
 
 #include "minishell.h"
 
-int	is_builtin(t_cmd *cmd)
-{
-	if (ft_strlen((*cmd->cmd)->content) == 2 && ft_strncmp((*cmd->cmd)->content, "cd", 2) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 4 && ft_strncmp((*cmd->cmd)->content, "echo", 4) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 3 && ft_strncmp((*cmd->cmd)->content, "env", 3) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 4 && ft_strncmp((*cmd->cmd)->content, "exit", 4) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 6 && ft_strncmp((*cmd->cmd)->content, "export", 6) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 3 && ft_strncmp((*cmd->cmd)->content, "pwd", 3) == 0)
-		return (1);
-	else if (ft_strlen((*cmd->cmd)->content) == 5 && ft_strncmp((*cmd->cmd)->content, "unset", 5) == 0)
-		return (1);
-	return (0);
-}
-
 int	exec_it(t_cmd *cmd, t_minishell *data)
 {
 	signal_default();
@@ -62,7 +43,7 @@ static int	launch_process(t_cmd *cmd, t_minishell *data)
 		if (open_n_dup(cmd, data->x) != 0)
 			ft_exit(data);
 		if (cmd->cmd)
-			exec_it(cmd , data);
+			exec_it(cmd, data);
 		else
 			ft_exit(data);
 	}
@@ -74,10 +55,9 @@ int	wait_child(t_xek *x)
 	int	i;
 	int	ret;
 
-	close_all(x);
-	i = 0;
+	i = -1;
 	ret = 0;
-	while (i < x->nb_cmd)
+	while (++i < x->nb_cmd)
 	{
 		if (!x->cmd[i].cmd || is_builtin(&x->cmd[i]) == 0)
 		{
@@ -88,19 +68,9 @@ int	wait_child(t_xek *x)
 				if (WIFEXITED(ret))
 					x->cmd->data->code_err = WEXITSTATUS(ret);
 				else if (WIFSIGNALED(ret))
-				{
-					signal_signal(ret);
-					x->cmd->data->code_err = 128 + WTERMSIG(ret);
-				}
-				else
-				{
-					x->cmd->data->code_err = WTERMSIG(ret) + 128;
-					if (x->cmd->data->code_err == 139)
-						err_write("Segmentation fault error\n", x->cmd->data->code_err);
-				}
+					handle_term_sig(ret);
 			}
 		}
-		i++;
 	}
 	return (0);
 }
@@ -109,10 +79,8 @@ int	wait_child(t_xek *x)
 int	go_exec(t_xek *x, t_minishell *data)
 {
 	int	i;
-	
-	/* On lance un process pour chaque commande */
+
 	i = 0;
-	// signal_ignore();
 	while (i < x->nb_cmd)
 	{
 		signal_ignore();
@@ -120,45 +88,11 @@ int	go_exec(t_xek *x, t_minishell *data)
 			launch_process(&(x->cmd[i]), data);
 		else if (is_builtin(&(x->cmd[i])) == 1)
 			if (handle_builtin(&(x->cmd[i]), data) != 0)
-				return (wait_child(x), 1);
+				return (close_all(x), wait_child(x), 1);
 		i++;
 	}
+	close_all(x);
 	wait_child(x);
-	// set_signal();
-	return (0);
-}
-
-int	open_n_leave(t_list	*token)
-{
-	int	fd;
-
-	while (token)
-	{
-		if (token->type == STDOUT)
-		{
-			fd = open(token->next->content, O_CREAT | O_TRUNC | O_RDWR, 0666);
-			close(fd);
-		}
-		else if (token->type == APPEND)
-		{
-			fd = open(token->next->content, O_CREAT | O_APPEND | O_RDWR, 0666);
-			close(fd);
-		}
-		else if (token->type == STDIN)
-		{
-			fd = open(token->content, O_RDONLY);
-			if (fd < 0)
-			{
-				put_str_err("minishell: ");
-				put_str_err(token->next->content);
-				put_str_err(": No such file or directory\n");
-				// if (!(x->nb_cmd > 1 && cmd->id == 0))
-				// 	cmd->data->code_err = 1;
-				return (1);
-			}
-		}
-		token = token->next;
-	}
 	return (0);
 }
