+++ sources/exec/exec_heredoc.c
@@ -6,28 +6,11 @@
 /*   By: jthuysba <jthuysba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/30 20:10:27 by jthuysba          #+#    #+#             */
-/*   Updated: 2023/07/05 13:20:50 by jthuysba         ###   ########.fr       */
+/*   Updated: 2023/07/05 16:28:30 by jthuysba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-/* Retourne le nombre de hdoc dans les tokens */
-int	nb_hdoc(t_minishell *data)
-{
-	t_list	*elem;
-	int		count;
-
-	elem = *(data->token);
-	count = 0;
-	while (elem)
-	{
-		if (elem->type == HEREDOC)
-			count++;
-		elem = elem->next;
-	}
-	return (count);
-}
 
 /* Recupere tous les delimiteurs des heredoc */
 static int	get_delims(t_minishell *data)
@@ -39,12 +22,11 @@ static int	get_delims(t_minishell *data)
 	i = 0;
 	while (elem)
 	{
-		/* Quand on trouve un heredoc on recup le delim
-		et on ouvre la pipe dans laquelle on va ecrire */
 		if (elem->type == HEREDOC)
 		{
 			data->x->hdoc[i].delim = ft_strdup(elem->next->content);
-			if (elem->next->quote_trace == 1  || check_dollar(data->x->hdoc[i].delim) == 1)
+			if (elem->next->quote_trace == 1
+				|| check_dollar(data->x->hdoc[i].delim) == 1)
 				data->x->hdoc[i].flag_sub = 1;
 			else
 				data->x->hdoc[i].flag_sub = 0;
@@ -57,49 +39,13 @@ static int	get_delims(t_minishell *data)
 	return (0);
 }
 
-/* Ecrit str dans fd (avec un \n) */
-static void write_in_fd(char *str, int fd)
-{
-	if (!str)
-		return ;
-	(void)write(fd, str, ft_strlen(str));
-	write(fd, "\n", 1);
-}
-
-char *substitute_hdoc(char *input, t_minishell *data)
-{
-	char			*result;
-	t_substitution	*s;
-	
-	s = malloc(sizeof(t_substitution));
-	ft_memset(s, 0, sizeof(t_substitution));
-	if (!input)
-		return (NULL);
-	while ((s->i <= ft_strlen(input)) && input[s->i])
-	{
-		sub_dollar_hdoc(s, input, data);
-	}
-	result = ft_strdup(s->new_content);
-	free(s->new_content);
-	free(input);
-	free(s);
-	return (result);
-}
-
-/* Ouvre l'ecriture du heredoc */
-static int	write_in_hdoc(t_hdoc *hdoc, t_minishell *data)
+static char	*read_n_write(t_hdoc *hdoc, t_minishell *data)
 {
 	char	*input;
 
-	handle_signal_hdoc();
 	input = readline("> ");
-	if (!input)
-	{
-		free(hdoc->delim);
-		close(hdoc->hd_pipe[0]);
-		close(hdoc->hd_pipe[1]);
-		return (1);
-	}
+	if (!input || ft_strcmp(input, hdoc->delim) == 0)
+		return (NULL);
 	if (input[0] == 0)
 		write_in_fd("", hdoc->hd_pipe[1]);
 	else
@@ -108,40 +54,32 @@ static int	write_in_hdoc(t_hdoc *hdoc, t_minishell *data)
 			input = substitute_hdoc(input, data);
 		write_in_fd(input, hdoc->hd_pipe[1]);
 	}
-	while (input && (!input[0] || ft_strcmp(input, hdoc->delim)) && (singleton_minishell()->code_err  != 130))
+	return (input);
+}
+
+/* Ouvre l'ecriture du heredoc */
+static int	write_in_hdoc(t_hdoc *hdoc, t_minishell *data)
+{
+	char	*input;
+
+	handle_signal_hdoc();
+	input = read_n_write(hdoc, data);
+	while (input && (!input[0] || ft_strcmp(input, hdoc->delim))
+		&& (singleton_minishell()->code_err != 130))
 	{
 		free(input);
-		if (singleton_minishell()->code_err  == 130)
+		if (singleton_minishell()->code_err == 130)
 			break ;
-		input = readline("> ");
-		if (!input || ft_strcmp(input, hdoc->delim) == 0)
-		{
-			free(hdoc->delim);
-			close(hdoc->hd_pipe[0]);
-			close(hdoc->hd_pipe[1]);
-			return (1);
-		}
-		if (input[0] == 0)
-			write_in_fd("", hdoc->hd_pipe[1]);
-		else
-		{
-			if (hdoc->flag_sub == 0)
-				input = substitute_hdoc(input, data);
-			write_in_fd(input, hdoc->hd_pipe[1]);
-		}
+		input = read_n_write(hdoc, data);
 	}
-	free(hdoc->delim);
-	free(input);
-	close(hdoc->hd_pipe[0]);
-	close(hdoc->hd_pipe[1]);
-	return (0);
+	return (free_hdoc_tools(hdoc, input), 0);
 }
 
 /* Child process dans lequel on va remplir le heredoc */
 static int	heredoc_child(t_minishell *data)
 {
 	int	i;
-	
+
 	i = 0;
 	while (i < data->x->nb_hdoc)
 	{
@@ -156,7 +94,7 @@ int	exec_heredoc(t_minishell *data)
 {
 	pid_t	pid;
 	int		i;
-	
+
 	data->x->nb_hdoc = nb_hdoc(data);
 	if (data->x->nb_hdoc == 0)
 		return (0);
