+++ sources/exec/handle_redir.c
@@ -6,7 +6,7 @@
 /*   By: jthuysba <jthuysba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/30 19:52:27 by jthuysba          #+#    #+#             */
-/*   Updated: 2023/07/03 13:45:16 by jthuysba         ###   ########.fr       */
+/*   Updated: 2023/07/05 17:26:00 by jthuysba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,17 @@ int	nb_redir(t_list	*elem)
 	return (count);
 }
 
+int	set_tools(t_cmd *cmd)
+{
+	cmd->redir = malloc(sizeof(int) * cmd->nb_redir);
+	if (!cmd->redir)
+		return (1);
+	cmd->file = ft_calloc(sizeof(char *) * (cmd->nb_redir + 1), sizeof(char *));
+	if (!cmd->file)
+		return (1);
+	return (0);
+}
+
 /* Enregistre les redirections avec leur type et leur file associes
 -> cat < file */
 int	handle_redir(t_cmd *cmd, t_list *elem)
@@ -37,13 +48,8 @@ int	handle_redir(t_cmd *cmd, t_list *elem)
 	cmd->nb_redir = nb_redir(*(cmd->token));
 	if (cmd->nb_redir == 0)
 		return (0);
-	cmd->redir = malloc(sizeof(int) * cmd->nb_redir);
-	if (!cmd->redir)
+	if (set_tools(cmd) != 0)
 		return (1);
-	cmd->file = ft_calloc(sizeof(char *) * (cmd->nb_redir + 1), sizeof(char *));
-	if (!cmd->file)
-		return (1);
-	// ft_memset(cmd->file, 0, sizeof(char *) * );
 	i = 0;
 	while (elem)
 	{
@@ -53,7 +59,7 @@ int	handle_redir(t_cmd *cmd, t_list *elem)
 			if (elem->next)
 				cmd->file[i] = ft_strdup(elem->next->content);
 			else
-				return (1);//WIP
+				return (1);
 			if (!cmd->file[i])
 				return (1);
 			i++;
@@ -61,6 +67,5 @@ int	handle_redir(t_cmd *cmd, t_list *elem)
 		elem = elem->next;
 	}
 	cmd->file[i] = 0;
-	// cmd->file[i] = ft_strdup(NULL);
 	return (0);
 }
